refactor: extract all turndown code into own module file
