--- scripts/utils/turndown.js
@@ -1,33 +1,0 @@
-const Turndown = require('turndown');
-
-const escapes = [
-  [/\\/g, '\\\\'],
-  [/\*/g, '\\*'],
-  [/^-/g, '\\-'],
-  [/^\+ /g, '\\+ '],
-  [/^(=+)/g, '\\$1'],
-  [/^(#{1,6}) /g, '\\$1 '],
-  [/`/g, '\\`'],
-  [/^~~~/g, '\\~~~'],
-  [/\[/g, '\\['],
-  [/\]/g, '\\]'],
-  [/_/g, '\\_'],
-  [/^(\d+)\. /g, '$1\\. '],
-
-  // Because we are converting to JSX, we need to ensure opening and closing < >
-  // are properly escaped to their HTML entities, otherwise it is parsed as the
-  // start of a component.
-  [/</g, '&lt;'],
-  [/>/g, '&gt;'],
-];
-
-// We need to override the built-in `escape` function so that we can use our
-// updated `escapes` variable from above, which differs slightly than the
-// default `escapes` that ships with turndown.
-Turndown.prototype.escape = (string) => {
-  return escapes.reduce((accumulator, escape) => {
-    return accumulator.replace(escape[0], escape[1]);
-  }, string);
-};
-
-module.exports = Turndown;
